+++ src/app/task/[type]/Calendar/Day.tsx
@@ -1,9 +1,4 @@
 "use client";
-
-import { Tooltip, TooltipContent } from "@/components/Tooltip";
-import { cn } from "@/lib/cn";
-import { TASK_COLOR } from "@/lib/consts";
-import { usePopupStore } from "@/stores/popup";
 import type {
   AppointmentFull,
   CalendarAppointment,
@@ -21,22 +16,12 @@ import moment from "moment";
 import { useSearchParams } from "next/navigation";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { useDrop } from "react-dnd";
-import { FaPen } from "react-icons/fa6";
-import { useMediaQuery } from "react-responsive";
 import { assignAppointmentDate } from "../../../../actions/appointment/assignAppointmentDate";
 import { updateTask } from "../../../../actions/task/dragTask";
 import mergeRefs from "merge-refs";
-import DraggableTaskTooltip from "./draggable/DraggableTaskTooltip";
-import {
-  formatDate,
-  formatTime,
-  updateTimeSpace,
-} from "@/utils/taskAndActivity";
-import { any } from "zod";
-import { ResizableBox } from "react-resizable";
-import "react-resizable/css/styles.css";
-import ResizeTaskTooltip from "./draggable/ResizeTaskTooltip";
-import zIndex from "@mui/material/styles/zIndex";
+import { formatTime, updateTimeSpace } from "@/utils/taskAndActivity";
+import DayRow from "../components/day/DayRow";
+import DayTask from "../components/day/DayTask";
 
 function useDate() {
   const searchParams = useSearchParams();
@@ -83,14 +68,8 @@ export default function Day({
   );
 
   const date = useDate();
-  // TODO:
-  const [boxHeight, setBoxHeight] = useState(200);
-  const { open } = usePopupStore();
-  const is1300 = useMediaQuery({ query: "(max-width: 1300px)" });
-  const [draggedOverRow, setDraggedOverRow] = useState<number | null>(null);
   const parentRef = useRef<HTMLDivElement>(null);
   const [isRefAvailable, setIsRefAvailable] = useState<boolean>(false);
-  const [isActiveResize, setIsActiveResize] = useState(false);
   const [{ canDrop, isOver }, dropRef] = useDrop({
     accept: ["tag", "task", "appointment"],
     collect: (monitor) => ({
@@ -286,82 +265,22 @@ export default function Day({
       ref={mergeRefs(dropRef, parentRef, containerRef)}
       className="relative mt-3 h-[90%] overflow-auto"
     >
-      {rows.map((row, i) => {
-        const rowTime = formatTime(row);
-        const dateRangeforBgChanger =
-          rowTime >= settings?.dayStart && rowTime <= settings?.dayEnd;
-
-        return (
-          <div key={i} className="relative">
-            <div
-              className={cn(
-                "absolute -top-[37.5px] flex h-full w-[100px] items-center justify-center text-[19px]",
-                i === 0 && "-top-5",
-              )}
-              style={{
-                color:
-                  rowTime >= settings?.dayStart && rowTime <= settings?.dayEnd
-                    ? "#7575a3"
-                    : "#d1d1e0",
-              }}
-            >
-              {row}
-            </div>
-            <button
-              type="button"
-              onDrop={(event: React.DragEvent) => {
-                handleDrop(event, i);
-                setDraggedOverRow(null);
-              }}
-              onDragOver={(event: React.DragEvent) => {
-                event.preventDefault();
-                setDraggedOverRow(i);
-              }}
-              onDragLeave={() => setDraggedOverRow(null)}
-              className={cn(
-                "ml-[85px] block h-[75px] border-neutral-200",
-                i !== rows.length && "border-b border-l",
-                i !== 0 ? "cursor-pointer" : "border-t",
-              )}
-              onClick={() => {
-                const date = formatDate(new Date());
-                const startTime = formatTime(row);
-                open("ADD_TASK", { date, startTime, companyUsers });
-              }}
-              disabled={i === 0}
-              style={{
-                backgroundColor:
-                  draggedOverRow === i
-                    ? "#c4c4c4"
-                    : dateRangeforBgChanger
-                      ? " white	"
-                      : "#f2f2f2",
-                width: "calc(100% - 85px)",
-              }}
-            >
-              {/* Row heading */}
-            </button>
-          </div>
-        );
-      })}
+      {rows.map((row, i) => (
+        <DayRow
+          key={i}
+          row={row}
+          rows={rows}
+          companyUsers={companyUsers}
+          index={i}
+          settings={settings}
+          onDrop={handleDrop}
+        />
+      ))}
 
       {/* Tasks */}
       {events.map((event, index) => {
-        const top = `${event.rowStartIndex * 75}px`;
-        // const height = `${(event.rowEndIndex - event.rowStartIndex + 1) * 55
-        //   }px`;
-        const widthNumber = is1300 ? 300 : 300;
-        let width = `${widthNumber}px`;
-        // @ts-ignore
-        const backgroundColor = event.priority
-          ? // @ts-ignore
-            TASK_COLOR[event.priority]
-          : "rgb(255, 255, 255)";
-
-        // Calculate how many tasks are in the same row
         const eventStartTime = moment(event.startTime, "HH:mm");
         const eventEndTime = moment(event.endTime, "HH:mm");
-        // sort by big indexes
         const tasksInRow = sortedEvents.filter((task) => {
           const taskStartTime = moment(task.startTime, "HH:mm");
           const taskEndTime = moment(task.endTime, "HH:mm");
@@ -373,139 +292,24 @@ export default function Day({
             return true;
           }
         });
-        const diffByMinutes = eventEndTime.diff(eventStartTime, "minutes");
-
-        const height = Math.round((diffByMinutes / 60) * 75); // 75 is the height of one task
-        // If there are more than one task in the same row
-        // then move the task right
-        // If there are more than two tasks in the same row
-        // then reduce the width of the task
-        // Now figure out which task is this
         const taskIndex = tasksInRow.findIndex((task) => task.id === event.id);
-        // If there's more than two tasks in the same row
-        // then reduce the width of the task
-        // calculate the width based on the number of tasks in the row (devide from 90%)
-        if (tasksInRow.length > 2) {
-          width = `${90 / tasksInRow.length}%`;
-        }
-
-        // Define a function to truncate the task title based on the height
-        const truncateTitle = (title: string, maxLength: number) => {
-          return title.length > maxLength
-            ? `${title.slice(0, maxLength)}...`
-            : title;
-        };
-
-        // Define the maximum title length based on the height
-        // const maxTitleLength =
-        //   height === "45px" ? 60 : height === "90px" ? 120 : event.title.length;
-        if (!isRefAvailable) return null;
         return (
-          <Tooltip key={event.id}>
-            <ResizeTaskTooltip
-              task={event}
-              height={height}
-              style={{
-                position: "absolute",
-                left: calculateLeftPosition(taskIndex, tasksInRow.length),
-                top,
-                backgroundColor,
-                borderRadius: "10px",
-              }}
-            >
-              <DraggableTaskTooltip
-                //@ts-ignore
-                className={`rounded-lg border-2 px-2 py-1 text-[17px] ${event.type === "appointment" ? "overflow-y-auto text-gray-600" : "text-white"} hover:z-20`}
-                style={{
-                  height: `${height}px`,
-                  maxWidth: width,
-                  minWidth: width,
-                }}
-                task={event}
-                updateTaskData={{ event, companyUsers }}
-                updateAppointmentData={{
-                  appointment: appointmentsFull.find(
-                    (appointment) => appointment.id === event.id,
-                  ),
-                  employees: companyUsers,
-                  customers,
-                  vehicles,
-                  templates,
-                  settings,
-                }}
-              >
-                {
-                  <>
-                    {event.type === "appointment" ? (
-                      <div className="flex h-full flex-col items-start text-xs">
-                        <div className="flex-1">
-                          <div className="flex items-center justify-between">
-                            <h3 className="text-sm font-semibold">
-                              {event.title}
-                            </h3>
-                          </div>
-                          <p className="text-left">
-                            Client:{" "}
-                            {event.client &&
-                              `${event.client.firstName} ${event.client.lastName}`}
-                          </p>
-                          <p className="text-left">
-                            Assigned To:{" "}
-                            {event.assignedUsers
-                              .slice(0, 1)
-                              .map(
-                                (user: User) =>
-                                  `${user.firstName} ${user.lastName}`,
-                              )}
-                          </p>
-                          <p className="text-left">
-                            {moment(event.startTime, "HH:mm").format("hh:mm A")}{" "}
-                            To{" "}
-                            {moment(event.endTime, "HH:mm").format("hh:mm A")}
-                          </p>
-                          <p className="text-left">
-                            Draft Estimate: {event.draftEstimate}
-                          </p>
-                        </div>
-                        <div className="absolute inset-y-1 right-0 h-[calc(100%-0.5rem)] w-1.5 rounded-lg border bg-[#6571FF]"></div>
-                      </div>
-                    ) : (
-                      <div className="flex h-full justify-start">
-                        <h3 className="font-semibold">{event.title}</h3>
-                      </div>
-                    )}
-                  </>
-                }
-              </DraggableTaskTooltip>
-              {event.type !== "appointment" && (
-                <TooltipContent className="w-72 rounded-md border border-slate-400 bg-white p-3">
-                  <div>
-                    <div className="flex items-center justify-between">
-                      <h3 className="font-semibold">{event.title}</h3>
-                      <button
-                        type="button"
-                        className="text- rounded-full bg-[#6571FF] p-2 text-white"
-                        onClick={() =>
-                          open("UPDATE_TASK", {
-                            task: event,
-                            companyUsers,
-                          })
-                        }
-                      >
-                        <FaPen className="mx-auto text-[10px]" />
-                      </button>
-                    </div>
-
-                    {/* @ts-ignore */}
-                    <p className="mt-3">{event.description}</p>
-
-                    {/* @ts-ignore */}
-                    <p className="mt-3">Task Priority: {event.priority}</p>
-                  </div>
-                </TooltipContent>
-              )}
-            </ResizeTaskTooltip>
-          </Tooltip>
+          <DayTask
+            key={event.id}
+            totalTaskInRow={tasksInRow.length}
+            calculateLeftPosition={calculateLeftPosition(
+              taskIndex,
+              tasksInRow.length,
+            )}
+            event={event}
+            appointmentsFull={appointmentsFull}
+            companyUsers={companyUsers}
+            isRefAvailable={isRefAvailable}
+            customers={customers}
+            settings={settings}
+            vehicles={vehicles}
+            templates={templates}
+          />
         );
       })}
     </div>
